refactor forcedTradeOffersToEntry to class
